+++ CHANGELOG.md
@@ -1,17 +1,23 @@
-# 1.0.0-beta (2024-07-06)
+# 1.0.0 (2024-07-15)
 
 
 ### Bug Fixes
 
 * gui, helper system bugs ([f4c7aa9](https://github.com/biyuehu/sirius/commit/f4c7aa9bf19d20d445349d9ca53ad2b73523fba1))
+* utils, gui, manger systems bug ([0baaca6](https://github.com/biyuehu/sirius/commit/0baaca65bb1ad353f6aa41374525525ea67aede6))
 
 
 ### Features
 
 * component plugin, import json ([aa40e9c](https://github.com/biyuehu/sirius/commit/aa40e9ce810c89382539b75357fe09551b5dcb77))
+* crypto and locale full ([3baf45a](https://github.com/biyuehu/sirius/commit/3baf45a84fb4e9a8773a466cce5e45fb6e5a1310))
 * gui, helper, teleport ([59c6543](https://github.com/biyuehu/sirius/commit/59c6543e90c3f563be0960caf841d810c0930aa4))
+* hunter, gui shop, land ([086dccb](https://github.com/biyuehu/sirius/commit/086dccb94a8ac25cdb0a0a5f648daaabb8ab0603))
+* land systems ([fa957dc](https://github.com/biyuehu/sirius/commit/fa957dc6a3c3cddcbfa6a41f64582ff21c8124ba))
+* moeny hunter,shop,safe mode ([36ccccb](https://github.com/biyuehu/sirius/commit/36ccccb2d0765f7c2cd53b9a11b752b45820f864))
 * pack, build dev and levilamina publish ([3cf9963](https://github.com/biyuehu/sirius/commit/3cf9963a5838f0e870be5142a92a788a2debc137))
 * teleport system ([239c31e](https://github.com/biyuehu/sirius/commit/239c31e929b773e88059552b685a66e7468a977e))
+* toml config parse and i18n ([5d257cb](https://github.com/biyuehu/sirius/commit/5d257cbc988a4787e9cec87645c26d95254ee29c))
 * utils, manger systems ([31970a3](https://github.com/biyuehu/sirius/commit/31970a393682c7c3cf2ba89387d571f7b6b2d148))